Removing trailing newline from run-results v6.json (#62)
In core we have a pre-commit action which ensures each file ends with a newline
character. If we commit our schema updates in core before copying them to this
repository, this repository recieves the newline character. Now in the grand scheme
of things, this shouldn't be problematic, and isn't in using this repository for
validating json. HOWEVER, dbt-core has a github workflow CI check which checks for
differences between this repositories schemas and schemas generated from the artifacts
in core. When that workflow in core generates the schemas from artifacts for comparison,
it doesn't format them, and so the generated schemas _never_ end with a newline during
the CI check. This means if the schemas we copy over from core happen _after_ committing
them to core, the CI check will _always_ fail. This is a problem and we should fix it by
either

1. use a formatter in this repository (schemas.getdbt.com) to always _remove_ EOF newline characters
2. use a formatter in this repository (schemas.getdbt.com) and the dbt-core workflow to always ensure
newline characters

However, this commit is not to fix that issue, but to fix the v6.json of run-results so that
the CI check will start passing in core.